Hz and dB are Doubles until written to stream
It's ok to lose information after calculations.

+++ src/main/kotlin/fi/evident/fab/proq2/BandGain.kt
@@ -18,6 +18,6 @@ internal class BandGain
 
     @Throws(IOException::class)
     fun write(writer: LittleEndianBinaryStreamWriter) {
-        writer.writeFloat(gain)
+        writer.writeFloat(gain.toFloat())
     }
 }
